+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Change Log
 
+- [Nov 10, 2024] Added support for including/excluding specific filenames using -I and -E options
 - [Sep 15, 2024] code-unpacker added
 - [Jun 22, 2024] max_depth option added
 - [May 01, 2024] initial commit